routing for journals wip
