--- src/routes/authentication/middleware/user.journal.post.ts
@@ -1,10 +1,0 @@
-export async function createNewJournal(
-  req: Express.Request,
-  res: Express.Response
-) {
-  try {
-    console.log("hi");
-  } catch (exception) {
-    console.log(exception);
-  }
-}
