+++ src/routes/journal/middleware/journal.validators.ts
@@ -1,0 +1,14 @@
+import { param } from "express-validator";
+import mongoose from "mongoose";
+
+export const mongooseJournalIdValidator = (): any[] => {
+  return [
+    param("journalId")
+      .exists()
+      .custom(async (value: string) => {
+        if (!mongoose.Types.ObjectId.isValid(value)) {
+          return Promise.reject("Journal id is invalid");
+        }
+      }),
+  ];
+};
