+++ src/server.ts
@@ -4,8 +4,9 @@ const cookieParser = require("cookie-parser");
 import express from "express";
 import cors from "cors";
 import connectDB from "./database.config";
-import authenticationRouter from "./routes/authentication/authentication";
-import userRouter from "./routes/user/user";
+import authenticationRouter from "./routes/authentication/authentication.route";
+import userRouter from "./routes/user/user.route";
+import journalRouter from "./routes/journal/journal.route";
 import { IS_PRODUCTION } from "./check-environment-variables";
 import { validateAPIKey } from "./routes/authentication/middleware/validate-api-key";
 
@@ -34,6 +35,7 @@ app.get("/", validateAPIKey, (_req, res) => {
 
 app.use("/api/auth", authenticationRouter);
 app.use("/api/user", userRouter);
+app.use("/api/journal", journalRouter);
 const port = app.get("port");
 
 const server = app.listen(port, () =>
